pyproject.toml
[project]
name = "rogue-scholar-api"
version = "0.17"
description = "API for the Rogue Scholar science blogging platform."
authors = [{ name = "Martin Fenner", email = "martin@front-matter.de" }]
requires-python = "~=3.13.0"
readme = "README.md"
license = "MIT"
keywords = ["science", "blogging", "quart", "api", "supabase", "inveniordm"]
dependencies = [
    "Quart>=0.19.2,<0.20",
    "tox>=4.11.3,<5",
    "hypercorn>=0.14.4,<0.15",
    "supabase~=2.9",
    "quart-db[postgresql]~=0.10.0",
    "python-dotenv>=1.0.0,<2",
    "quart-schema>=0.17.1,<0.18",
    "pydash~=7.0",
    "quart-rate-limiter>=0.10.0,<0.11",
    "feedparser>=6.0.10,<7",
    "furl>=2.1.3,<3",
    "beautifulsoup4>=4.12.2,<5",
    "xmltodict>=0.12.0,<0.13",
    "nh3>=0.2.14,<0.3",
    "langdetect>=1.0.9,<2",
    "dateutils>=0.6.12,<0.7",
    "lxml>=5.1.0,<6",
    "pandoc~=2.4",
    "commonmeta-py>=0.201,<1",
    "python-frontmatter~=1.1",
    "babel>=2.14.0,<3",
    "iso8601>=2.1.0,<3",
    "levenshtein>=0.26.0,<0.28",
    "quart-cors>=0.7.0,<0.8",
    "httpx>=0.28,<0.30",
    "orjson>=3.9.14,<4",
    "weasyprint~=67.0",
    "setuptools>=0.70",
    "backoff>=2.2.1,<3",
]

[project.urls]
Homepage = "https://rogue-scholar.org"
Repository = "https://github.com/front-matter/rogue-scholar-api"
Documentation = "https://docs.rogue-scholar.org"

[project.scripts]
start = "api:run"

[dependency-groups]
dev = [
    "coverage",
    "pytest~=8.3",
    "pytest-cov~=4.1",
    "pytest-asyncio>=0.24,<0.25",
    "pytest-recording>=0.13,<0.14",
    "vcrpy~=5.1",
    "ruff>=0.7.0,<0.8",
]

[tool.hatch.build.targets.sdist]
include = ["api"]

[tool.hatch.build.targets.wheel]
include = ["api"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
select = ["F401", "F403"]

[tool.tox]
legacy_tox_ini = """
    [tox]
    min_version = 4.0
    env_list =
        py312

    [testenv]
    passenv = *
    deps =
        pytest
        pytest-cov
    commands = pytest --cov=rogue_scholar_api --cov-report=xml

    [coverage:run]
    relative_files = True
    branch = True
"""

[tool.pytest.ini_options]
asyncio_mode = "auto"
